lora_task/requirements.txt
torch>=2.1
torchvision
accelerate>=0.23.0
diffusers>=0.25.0
transformers>=4.33
peft>=0.7.0
pillow
safetensors
tqdm 